lightbox cleanup
