+++ src/Lightbox/Lightbox.tsx
@@ -1,51 +1,31 @@
 import React from 'react';
-import { Box } from '../Box';
-// import { Modal, ModalCloseButton, ModalContent, ModalOverlay } from '../Modal';
-// import { useTheme } from '../ThemeProvider';
-// import { SlideIn } from '../Transition';
-// // import { LightboxGalleryControls } from './components';
-import { LightboxProps } from './types';
+import ReactLightbox from 'react-image-lightbox';
+import { useGalleryContext } from './components';
 
-export const Lightbox = (props: LightboxProps) => {
-    return <Box />;
-    //     const { zIndices } = useTheme();
-    //     const { isOpen, onClose, showControls, onKeyDown, children } = props;
+// still uses context and self-adding images
+// Switch to use lightbox library - no need to reinvent the wheel
+export const LightboxGallery = () => {
+    const context = useGalleryContext();
+    const { activeItem, activeIndex, media, setActiveItem, onPrev, onNext } = context;
 
-    //     return (
-    //         <Box>
-    //             <SlideIn offset="10px" in={isOpen}>
-    //                 {/* @ts-ignore */}
-    //                 {styles => (
-    //                     <Modal
-    //                         isOpen={isOpen}
-    //                         onClose={onClose}
-    //                         preserveScrollBarGap
-    //                         isCentered
-    //                         blockScrollOnMount
-    //                         size="100%"
-    //                     >
-    //                         {/* solid background ModalOverlay */}
-    //                         <ModalOverlay opacity={1} backgroundColor="rgba(11,11,11,0.9)" />
+    if (activeItem && media?.length > 0) {
+        const current = media[activeIndex]?.src;
+        const next = media[(activeIndex + 1) % media.length]?.src;
+        const prev = media[(activeIndex + media.length - 1) % media.length]?.src;
+        return (
+            <ReactLightbox
+                mainSrc={current}
+                // mainSrcThumbnail={current}
+                nextSrc={next}
+                // nextSrcThumbnail={next}
+                prevSrc={prev}
+                // prevSrcThumbnail={prev}
+                onCloseRequest={() => setActiveItem(null)}
+                onMovePrevRequest={() => onPrev()}
+                onMoveNextRequest={() => onNext()}
+            />
+        );
+    }
 
-    //                         {/* overlay wrapper for close & navigation button actions */}
-    //                         <ModalOverlay bg="transparent" onKeyDown={onKeyDown}>
-    //                             <ModalCloseButton zIndex={zIndices.modal + 1} color="white" />
-    //                             {showControls && <LightboxGalleryControls />}
-    //                             <ModalContent
-    //                                 {...styles}
-    //                                 shadow={0}
-    //                                 bg="transparent"
-    //                                 h="100%"
-    //                                 my={0}
-    //                                 // pt="48px"
-    //                                 // pb="sm"
-    //                             >
-    //                                 {children}
-    //                             </ModalContent>
-    //                         </ModalOverlay>
-    //                     </Modal>
-    //                 )}
-    //             </SlideIn>
-    //         </Box>
-    //     );
+    return null;
 };
